Pre Final Update

+++ Package_Procedures.sql
@@ -1,25 +1,27 @@
 SET SERVEROUTPUT ON;
 SET LINESIZE 32000;
 
-
+column id format 99
+column name format a20
+column age format 99
+column gender format a10
+column religion format a10
+column height format 9.9
+column occupation format a10
+column home_Town format a10
+column language format a10
 
 
 CREATE OR REPLACE PACKAGE mypack 
 AS 
-	PROCEDURE Insert_Entry(name in varchar2,age in integer,gender in varchar2,religion in varchar2,height in float,occupation in varchar2,home_Town in varchar2,language in varchar2);
+	PROCEDURE Insert_Entry(name in varchar2,age in integer,gender in varchar2,religion in varchar2,height in number,occupation in varchar2,home_Town in varchar2,language in varchar2);
 	PROCEDURE Delete_Entry(B IN NUMBER);
 	PROCEDURE Search_Entry(Options in number);
-	PROCEDURE Upd(A in integer,name1 in varchar2,age1 in integer,gender1 in varchar2,religion1 in varchar2,height1 in float,occupation1 in varchar2,home_Town1 in varchar2,language1 in varchar2);
+	PROCEDURE Upd(A in integer,name1 in varchar2,age1 in integer,gender1 in varchar2,religion1 in varchar2,height1 in number,occupation1 in varchar2,home_Town1 in varchar2,language1 in varchar2);
 	FUNCTION  Show RETURN NUMBER;
 	
 END mypack; 
 /
-
-
-
-
-
-
 
 
 CREATE OR REPLACE PACKAGE BODY mypack 
@@ -27,7 +29,7 @@ AS
 
 	--Insert Entry
 
-	PROCEDURE Insert_Entry(name in varchar2,age in integer,gender in varchar2,religion in varchar2,height in float,occupation in varchar2,home_Town in varchar2,language in varchar2) IS
+	PROCEDURE Insert_Entry(name in varchar2,age in integer,gender in varchar2,religion in varchar2,height in number,occupation in varchar2,home_Town in varchar2,language in varchar2) IS
 	
 	
 	A number;
@@ -65,6 +67,7 @@ AS
 			maximum:=maximum+1;
 				
 			IF maximum<=10 THEN
+				
 				insert into data1 values(maximum,name,age,gender,religion,height,occupation,home_Town,language);
 				commit;
 			
@@ -100,19 +103,36 @@ AS
 	
 	PROCEDURE Delete_Entry(B IN NUMBER) IS
 	
+	rowExist NUMBER := 0;
 	BEGIN
 		IF B>=1 AND B<=10 THEN
-			DELETE FROM data1 WHERE id=B;
-			commit;
-	
+			select id into rowExist from data1 where id=B;
+			
+			IF rowExist is not null THEN
+				DELETE FROM data1 WHERE id=B;
+				commit;
+				DBMS_OUTPUT.PUT_LINE('Successfully Deleted ID:' ||B);
+				rowExist:=0;
+			END IF;
+
 		END IF;
 	
 		IF B>10 THEN
-			DELETE FROM data1@site1 WHERE id=B;
-			commit;
-	
+			select id into rowExist from data1@site1 where id=B;
+			
+			IF rowExist is not null THEN
+			
+				DELETE FROM data1@site1 WHERE id=B;
+				commit;
+				DBMS_OUTPUT.PUT_LINE('Successfully Deleted ID:' ||B);
+
+			END IF;
+		
 		END IF;
 		
+		EXCEPTION
+			WHEN no_data_found THEN
+				DBMS_OUTPUT.PUT_LINE('Not Valid ID');
 		
 	END Delete_Entry;
 	
@@ -120,25 +140,52 @@ AS
 	--UPDATING value
 	
 	
-	PROCEDURE Upd(A in integer,name1 in varchar2,age1 in integer,gender1 in varchar2,religion1 in varchar2,height1 in float,occupation1 in varchar2,home_Town1 in varchar2,language1 in varchar2) IS
+	PROCEDURE Upd(A in integer,name1 in varchar2,age1 in integer,gender1 in varchar2,religion1 in varchar2,height1 in number,occupation1 in varchar2,home_Town1 in varchar2,language1 in varchar2) IS
+	
+	idExist NUMBER := 0;
+	
 	BEGIN
 		
 		IF A<=10 THEN
-			--DBMS_OUTPUT.PUT_LINE('The Name is' ||name1);
-			UPDATE data1 SET name=name1,age=age1,gender=gender1,religion=religion1,height=height1,occupation=occupation1,home_Town=home_Town1,language=language1 WHERE id=A;
-			commit;
+			select id into idExist from data1 where id=A;
+			
+			IF idExist is not null THEN
+				
+				UPDATE data1 SET name=name1,age=age1,gender=gender1,religion=religion1,height=height1,occupation=occupation1,home_Town=home_Town1,language=language1 WHERE id=A;
+				commit;
+				idExist:=0;
+			END IF;
+			
 		ELSIF A>10 THEN
-			--DBMS_OUTPUT.PUT_LINE('The Name is' ||name1);
-			UPDATE data1@site1 SET name=name1,age=age1,gender=gender1,religion=religion1,height=height1,occupation=occupation1,home_Town=home_Town1,language=language1 WHERE id=A;
-			commit;
+			select id into idExist from data1@site1 where id=A;
+			
+			IF idExist is not null THEN
+				
+				UPDATE data1@site1 SET name=name1,age=age1,gender=gender1,religion=religion1,height=height1,occupation=occupation1,home_Town=home_Town1,language=language1 WHERE id=A;
+				DBMS_OUTPUT.PUT_LINE('Successfully Updated Values For ID:' ||A);
+				commit;
+				idExist:=0;
+			END IF;
 		
 		END IF;
-	
+		
+	EXCEPTION
+      WHEN DUP_VAL_ON_INDEX THEN
+         DELETE FROM data1 WHERE id=A;
+		 
+		 insert into data1 values(A,name1,age1,gender1,religion1,height1,occupation1,home_Town1,language1);
+		 commit;
+	  WHEN no_data_found THEN
+				DBMS_OUTPUT.PUT_LINE('Not Valid ID');
 	END Upd;
 	
 	
 	
+	
+	
 	--Searching 
+	
+	
 	
 	PROCEDURE Search_Entry(Options in number) IS
 	BEGIN
@@ -232,8 +279,44 @@ AS
 		
 			END LOOP;
 			
-			
 		ELSIF Options=8 THEN
+			
+			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Occupation(Actor):'||CHR(10));
+			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			FOR R IN (select id,name,age,gender,religion,occupation,home_Town,height,language from data1 where occupation='Actor' union select id,name,age,gender,occupation,religion,home_Town,height,language from data1@site1 where occupation='Actor' order by id) LOOP
+	
+				DBMS_OUTPUT.PUT_LINE('ID:'||R.id||' || ' ||'Name:'||R.name||' || ' || ' Age:'||R.age ||' || ' || ' Gender:' ||R.gender ||' || ' || ' Religion:' ||R.religion ||' || ' || ' Occupation:' ||R.occupation||' || ' || ' Height:' ||R.height ||' || ' || ' HomeTown:' ||R.home_Town ||' || ' || ' Language:' ||R.language); 
+				DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			END LOOP;
+		
+		ELSIF Options=9 THEN
+			
+			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Occupation(Actress):'||CHR(10));
+			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			FOR R IN (select id,name,age,gender,religion,occupation,home_Town,height,language from data1 where occupation='Actress' union select id,name,age,gender,occupation,religion,home_Town,height,language from data1@site1 where occupation='Actress' order by id) LOOP
+	
+				DBMS_OUTPUT.PUT_LINE('ID:'||R.id||' || ' ||'Name:'||R.name||' || ' || ' Age:'||R.age ||' || ' || ' Gender:' ||R.gender ||' || ' || ' Religion:' ||R.religion ||' || ' || ' Occupation:' ||R.occupation||' || ' || ' Height:' ||R.height ||' || ' || ' HomeTown:' ||R.home_Town ||' || ' || ' Language:' ||R.language); 
+				DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			END LOOP;
+			
+		ELSIF Options=10 THEN
+			
+			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Occupation(Teacher):'||CHR(10));
+			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			FOR R IN (select id,name,age,gender,religion,occupation,home_Town,height,language from data1 where occupation='Teacher' union select id,name,age,gender,occupation,religion,home_Town,height,language from data1@site1 where occupation='Teacher' order by id) LOOP
+	
+				DBMS_OUTPUT.PUT_LINE('ID:'||R.id||' || ' ||'Name:'||R.name||' || ' || ' Age:'||R.age ||' || ' || ' Gender:' ||R.gender ||' || ' || ' Religion:' ||R.religion ||' || ' || ' Occupation:' ||R.occupation||' || ' || ' Height:' ||R.height ||' || ' || ' HomeTown:' ||R.home_Town ||' || ' || ' Language:' ||R.language); 
+				DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
+		
+			END LOOP;
+			
+			
+		ELSIF Options=11 THEN
 			
 			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Height=5.2-5.5 and Gender=Female:'||CHR(10));
 			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
@@ -245,7 +328,7 @@ AS
 		
 			END LOOP;
 			
-		ELSIF Options=9 THEN
+		ELSIF Options=12 THEN
 			
 			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Height=5.2"-5.11" and Gender=Male:'||CHR(10));
 			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
@@ -258,7 +341,7 @@ AS
 			END LOOP;
 			
 			
-		ELSIF Options=10 THEN
+		ELSIF Options=13 THEN
 			
 			DBMS_OUTPUT.PUT_LINE(CHR(10) ||'Search By Religion=Islam and Gender=Female and Occupation=Engineer:'||CHR(10));
 			DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
@@ -276,10 +359,12 @@ AS
 	END Search_Entry;
 	
 	
-		--Show Function	
+	
+	
+	--Show Function	
 	
 	FUNCTION show
-		RETURN NUMBER AS
+	RETURN NUMBER AS
 
 	BEGIN
 		DBMS_OUTPUT.PUT_LINE(CHR(10));
@@ -289,7 +374,7 @@ AS
 		
 		DBMS_OUTPUT.PUT_LINE('ID:'||R.id||' || ' ||'Name:'||R.name||' || ' || ' Age:'||R.age ||' || ' || ' Gender:' ||R.gender ||' || ' || ' Religion:' ||R.religion ||' || ' || ' Occupation:' ||R.occupation||' || ' || ' Height:' ||R.height ||' || ' || ' HomeTown:' ||R.home_Town ||' || ' || ' Language:' ||R.language); 
 		DBMS_OUTPUT.PUT_LINE('------------------------------------------------------------------------------------------------------------------------------------------------');
-		--DBMS_OUTPUT.PUT_LINE(CHR(10));
+	
 		END LOOP;
 		
 
@@ -304,16 +389,7 @@ AS
 	END show;
 	
 	
-			
 
-
-		
-		
-
-	
-	
-	
-	
 END mypack; 
 /
